Added ClickTaskControllerTest.java and ClickTaskPresenterTest.java in the test package.
